Revert "Update on movement mechanics, not entirely finished vol.2"
