+++ Assets/Prefabs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-
 guid: bc0e035c6b2bddc4c8a7c168f67fc757
-
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
