--- Assets/Prefabs/Platform.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 078753bda87ee2d4782340080c057a3f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
